requirements.txt
apache-airflow
kubernetes
pandas